# Seed HashMaps thread-locally, straight from the OS.