--- a/src/libstd/lib.rs
+++ b/src/libstd/lib.rs
@@ -436,6 +436,7 @@ mod memchr;
 pub mod rt;
 mod panicking;
 mod rand;
+mod internal_rand;
 
 // Some external utilities of the standard library rely on randomness (aka
 // rustc_back::TempDir and tests) and need a way to get at the OS rng we've got