@huonw 2016-02-02
cc @rust-lang/libs 

> This reduces how much rand is required in std

I only had a few minutes to write this, and wanted to get the conversation started if there is one. The next step will be a patch that moves the rest of the private `std::rand` out of `std` (there's some fiddly bits that may take a few compile-edit cycles, since various tests use it, which is why I haven't done it here).