@brson 2016-02-02
> This new scheme has the slight downside of being consistent on a single
> thread

Can you expand on what this means and why it is? I don't see why changing the source of randomness would affect the behavior.