@Gankro 2016-02-02
@brson today HashMap::new is like:

`get_random_seed_from_os()`

with the PR it's like

`rngs.entry(thread_id).or_insert_with(get_random_seed_from_os)`

Today if you make two HashMaps and perform identical operations on them, you'll get the different iteration orders. With this patch, they will iterate identically. Historically programmers have been all-too-happy to latch onto any perceived determinism in HashMap iteration order. This is why JavaScript now  specifies that an object is semantically a LinkedHashMap. People relied on this fact, and when it didn't hold drop-downs got shuffled.

Now, this is less of a risk for Rust, because the JS case is relying on individual maps being *really* reliable, and this doesn't hold today or with this patch. However, one could imagine something to the effect of:

```rust
fn test() {
   let results: Vec<Vec<u8>> = vec![];
   for vec in get_data_sets() {
      let map = HashMap::new();
      for k in vec {
        map.insert(k, compute(k));
      }
      results.push(map.values().cloned().collect()); 
   }
   assert!(results[0] == results[1]);
}
```

Today this would be hopelessly broken, but with this patch this will "happen to work" if the input/output sequences line up. This is one of those situations where we can wag our finger and tell people not to do it, but at the end of the day it's externally observable it can and will be leveraged.

Note that the following will still be hopelessly and obviously broken:

```rust
fn test() {
  // I got this when I ran it the first time, shouldn't change!
  let expected = vec![1, 3, 5, 6, 11, 13];
  let map = HashMap::new();
  for k in get_test_input() {
    map.insert(k, compute(k));
  }
  let result = map.values().cloned().collect();
  assert_eq!(result, expected);
}
```

Which is the thing I would most worry about.