@alexcrichton 2016-02-02
Wow those are some impressive numbers! In the past I don't remember them being quite so dramatic...

If we're moving to thread-locals I wouldn't mind moving to full-blown process-globals. We've concluded in the past that there's no need in terms of DOS protection to have per-hashmap or even per-thread keys, a global one should be good enough.