@huonw 2016-02-02
> Can you expand on what this means and why it is? I don't see why changing the source of randomness would affect the behavior.

Currently each hashmap has different keys since they retrieve new ones from the thread local on creation. This patch changes those `u64`s  to be generated once and cached per-thread so all maps created on a single thread have the same keys.

> Wow those are some impressive numbers! In the past I don't remember them being quite so dramatic...

It'll probably depend on the platform/Linux kernel version, since they'll change the performance characteristics of retrieving random numbers (the old thread rng requires retrieving kilobytes, while this just needs a few).

> If we're moving to thread-locals I wouldn't mind moving to full-blown process-globals. We've concluded in the past that there's no need in terms of DOS protection to have per-hashmap or even per-thread keys, a global one should be good enough.

Yeah, definitely possible, but this implementation is particularly simple (don't have to worry about CASing anything etc). Very happy to switch to that if people prefer it.