@ranma42 2016-02-02
It would be possible to address @Gankro's concern by generating the actual keys with a per-thread PRNG seeded with the OS random data.