@ranma42 2016-02-02
The improvement from 130_000 ns to 3000 ns looks huge. Would it make sense to investigate why `thread_rng` is so slow (and maybe backport these changes to it)?