@huonw 2016-02-02
> EDIT: but isn't this what the old implementation was already doing?

Yes, it was, and it's why I was sure to point out this change has that downside in the PR. If we absolutely want to proactively help people avoid relying on hashmap order, a simpler and more efficient method would be to just increment one of the cached keys each time. (NB. this doesn't work quite so well with global rather than thread-local values: the locked increment/contention makes things slower.)

> Would it make sense to investigate why thread_rng is so slow (and maybe backport these changes to it)?

Maybe, but it's fundamental to the algorithm (it requires an extensive initialisation). In any case, the thread_rng in std is only used here, and hence is being removed.