@ranma42 2016-02-02
If the update of the global cached keys is that simple, we could do atomic increments. There might be some contention, but it should still be way faster than locking.

Another option would be to keep the thread-local approach and use a simpler (non crypto-safe) RNG to have cheaper initialisation and key generation. In the repo for the `rand` crate there has been some [discussion](https://github.com/rust-lang-nursery/rand/pull/59) about providing fast weak RNG, which might be a good fit for this purpose: they would be marginally slower than the no-op key copy and still shuffle the keys in a different way for each `HashMap`