@huonw 2016-02-03
(Incidentally, my method of measuring had a lot of overhead due to the timers, I've updated the numbers.)

---

> If the update of the global cached keys is that simple, we could do atomic increments. There might be some contention, but it should still be way faster than locking.

Yes, but... locking was never proposed? In the single threaded case (i.e. no contention), doing the atomic increment seems to be more than 2.8 times slower than the non-incrementing global version and both thread local ones (incrementing or not): it's about 5.4 ns / invocation.

> Another option would be to keep the thread-local approach and use a simpler (non crypto-safe) RNG to have cheaper initialisation and key generation. In the repo for the rand crate there has been some discussion about providing fast weak RNG, which might be a good fit for this purpose: they would be marginally slower than the no-op key copy and still shuffle the keys in a different way for each HashMap

One could say that the incrementing was exactly this sort of non-crypto safe RNG (and is about as fast as you can get). There's two factors at play here:

- the key for a hashmap should be hard to deduce externally to protect against HashDoS
- the order of keys within a hashmap is undefined, and hence people shouldn't accidentally rely on it (to ensure their code doesn't break when things change slightly)

The first we are assuming (see e.g. #27243) is guaranteed by the combination of 128-bits of randomness plus the design of SipHash, so this discussion is all about the second.