@brson 2016-02-28
I'm still a fan of this. @huonw is this ready to go?