@alexcrichton 2016-03-03
The libs team discussed this during triage yesterday and the conclusion was:

* The speedup wins here seem quite good, so we should land
* It would be nice for now to preserve nondeterministic iteration order among all hash maps
* We can probably recover this via `Cell<u64>` in TLS and it's just bumped each time we hit it.

Once that's in place this should be good to go, thanks @huonw!