@pczarn 2016-03-24
I suggest preserving nondeterministic iteration order only in debug builds, not release builds. However, there's no rush to change that.

(It seems the only flag for conditional compilation in debug builds is `cfg!(debug_assertions)`.)